lint .test.sol

+++ contracts/v1/test/InstaAccountV3.test.sol
@@ -7,12 +7,9 @@ pragma experimental ABIEncoderV2;
  */
 
 contract Record {
-    uint public constant version = 3;
-
+    uint256 public constant version = 3;
 }
 
 contract InstaAccountV3 is Record {
-
     receive() external payable {}
-
 }
